save work
git-svn-id: http://jmonkeyengine.googlecode.com/svn/trunk@2156 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ src/com/jme/sound/openAL/objects/Sample3D.java
@@ -137,7 +137,7 @@ public class Sample3D extends SoundSpatial{
         
     }
     
-    public void setVolume(int volume){
+    public void setVolume(float volume){
         AL10.alSourcef(sourceNumber, AL10.AL_GAIN, volume);
     }
